+++ tests/test_data_api_client.py
@@ -715,6 +715,19 @@ class TestSupplierMethods(object):
         assert rmock.called
         assert rmock.request_history[0].json() == {'updated_by': 'supplier'}
 
+    def test_reset_dos_framework_contact_information(self, data_client, rmock):
+        rmock.post(
+            'http://baseurl/suppliers/123/framework-contact-information/reset-dos-contact',
+            json={'suppliers': 'result'},
+            status_code=201,
+        )
+
+        result = data_client.reset_dos_framework_contact_information(123, 'supplier')
+
+        assert result == {'suppliers': 'result'}
+        assert rmock.called
+        assert rmock.request_history[0].json() == {'updated_by': 'supplier'}
+
     def test_update_contact_information(self, data_client, rmock):
         rmock.post(
             'http://baseurl/suppliers/123/contact-information/2',